+++ module-api/src/main/java/com/mile/controller/user/UserController.java
@@ -4,9 +4,9 @@ import com.mile.dto.SuccessResponse;
 import com.mile.exception.message.SuccessMessage;
 import com.mile.external.client.dto.UserLoginRequest;
 import com.mile.token.service.TokenService;
-import com.mile.user.serivce.UserService;
-import com.mile.user.serivce.dto.AccessTokenGetSuccess;
-import com.mile.user.serivce.dto.LoginSuccessResponse;
+import com.mile.user.service.UserService;
+import com.mile.user.service.dto.AccessTokenGetSuccess;
+import com.mile.user.service.dto.LoginSuccessResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
