+++ module-api/src/main/java/com/mile/controller/user/UserControllerSwagger.java
@@ -3,8 +3,8 @@ package com.mile.controller.user;
 import com.mile.dto.ErrorResponse;
 import com.mile.dto.SuccessResponse;
 import com.mile.external.client.dto.UserLoginRequest;
-import com.mile.user.serivce.dto.AccessTokenGetSuccess;
-import com.mile.user.serivce.dto.LoginSuccessResponse;
+import com.mile.user.service.dto.AccessTokenGetSuccess;
+import com.mile.user.service.dto.LoginSuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
