+++ module-domain/src/main/java/com/mile/user/service/dto/AccessTokenGetSuccess.java
@@ -1,4 +1,4 @@
-package com.mile.user.serivce.dto;
+package com.mile.user.service.dto;
 
 public record AccessTokenGetSuccess(
         String accessToken
